explicitly use "binary" encoding to update salt
Must be a change in Node.js, but the test for slapdpasswd was now
failing for no reason.

Mostly just random guessing, but added 'binary' as explicit encoding for
the ctx.update() call when adding salt to the crypto and the test now
passes.

Must be the complete randomness coming out of slapdpasswd generates
strange characters, compared with other tests that just use "salt" etc

Should probably add a few more of these?  Maybe one with gibberish
passwords?